+++ Parsing/global.h
@@ -1,0 +1,13 @@
+#ifndef GLOBAL_H
+#define GLOBAL_H
+#define MAX_LINES 1000
+#define MAX_TOKENS 3
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
+#include <stdbool.h>
+#include "lexical.h"
+#include "parser.h"
+#include "vm.h"
+#endif
