+++ Parsing/parser.h
@@ -86,44 +86,3 @@ void freeAST(astNode_t *node) {
     free(node->children);
     free(node);
 }
-
-void labelLastChild(astNode_t* node)
-{
-
-}
-
-// Traverse the AST to check the syntax
-void checkSyntax(astNode_t* node) 
-{
-    // ************************************************************************
-    // current parent token type
-    bool inRoot = false;
-    bool inLabel = false;
-    bool inInstruction = false;
-    // *************************************************************************
-    // If the node is NULL, return
-    if (node == NULL) 
-    {
-        return;
-    }
-    // Check the syntax of current node and of the children of the current node
-    for(int i = 0; i < node->numChildren; i++)
-    {
-        if(strcmp(node->token.type,"ROOT")==0)
-        {
-            inRoot = true;
-        }
-        else if(strcmp(node->token.type,"LABEL")==0)
-        {
-            inLabel = true;
-        }
-        else if(strcmp(node->token.type,"INSTRUCTION")==0)
-        {
-            inInstruction = true;
-        }
-    }
-    for(int i = 0; i < node->numChildren; i++)
-    {
-        checkSyntax(node->children[i]);
-    }
-}
