--- Parsing/prog.exe.stackdump
@@ -1,40 +1,0 @@
-Stack trace:
-Frame         Function      Args
-0007FFF54C40  0001800602E4 (0007FFF54E48, 0007FFFFCE00, 000000000002, 0007FFFFDC10) msys-2.0.dll+0x202E4
-FFFFFFFFFFFEFEDF  0001800622F5 (0007FFF555F0, 000000000000, 0000000001F8, 000000000000) msys-2.0.dll+0x222F5
-0007FFF55350  0001800C0D44 (000000000000, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x80D44
-000000000006  0001800BC819 (000000000000, D85D078A19D7, 000000000000, 00000000002E) msys-2.0.dll+0x7C819
-0007FFF557A0  0001800BCC1A (0007FFF55670, 000000000000, 0001800AD555, 000A00001A70) msys-2.0.dll+0x7CC1A
-0007FFF557A0  0001802124A8 (000000000000, 000A00017028, 000100401626, 000000000000) msys-2.0.dll+0x1D24A8
-0007FFF557A0  0001802125D5 (000100401F91, 0007FFFBDBD8, 0007FFF56D64, 000000000000) msys-2.0.dll+0x1D25D5
-0007FFF557A0  00018017826F (000A000003A0, 000300000000, 003000000010, 0007FFF55830) msys-2.0.dll+0x13826F
-0007FFF557A0  00018019331B (000A000003A0, 000300000000, 003000000010, 0007FFF55830) msys-2.0.dll+0x15331B
-0007FFF557A0  0001004023FB (000700000000, 0007FFF557D0, 000000000000, 5443555254534E49) prog.exe+0x23FB
-0007FFF55830  000100402568 (000A00000400, 0007FFFFCE00, 0000800455E0, 000700000000) prog.exe+0x2568
-0007FFFFCD30  000180047ED1 (000000000000, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x7ED1
-000000000000  000180045AC3 (000000000000, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x5AC3
-0007FFFFFFF0  000180045B74 (000000000000, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x5B74
-End of stack trace
-Loaded modules:
-000100400000 prog.exe
-7FFF62BE0000 ntdll.dll
-7FFF61770000 KERNEL32.DLL
-7FFF606A0000 KERNELBASE.dll
-000180040000 msys-2.0.dll
-7FFF62A00000 advapi32.dll
-7FFF62940000 msvcrt.dll
-7FFF616D0000 sechost.dll
-7FFF60C20000 RPCRT4.dll
-7FFF5F730000 CRYPTBASE.DLL
-7FFF60460000 bcryptPrimitives.dll
-7FFF54180000 netapi32.dll
-7FFF56100000 SAMCLI.DLL
-7FFF604E0000 ucrtbase.dll
-7FFF53E10000 SAMLIB.dll
-7FFF5EB90000 NETUTILS.DLL
-7FFF61CB0000 user32.dll
-7FFF60190000 win32u.dll
-7FFF622C0000 GDI32.dll
-7FFF60070000 gdi32full.dll
-7FFF60600000 msvcp_win.dll
-7FFF62410000 IMM32.DLL
